#6.14 Locking Down DeleteItem Permissions

+++ backend/src/resolvers/Mutation.js
@@ -46,9 +46,15 @@ const Mutations = {
   deleteItem: async (parent, args, ctx, info) => {
     const where = { id: args.id };
     // 1.find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user{ id }}`);
     //2. check if they own that itme, or have the permissions
-    //TODO:
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+    if (!ownsItem && !hasPermission) {
+      throw new Error("You don't have the permission to do that!");
+    }
     //3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -139,21 +145,24 @@ const Mutations = {
     return updatedUser;
   },
   updatePermissions: async (parent, args, ctx, info) => {
-    console.log("HEllo");
+    console.log("[MUTATION] updatePermissions");
     // 1. Check if they are logged in
     if (!ctx.request.userId) throw new Error("You must be logged in!");
     // 2. Query the current user
-    const currentUser = await ctx.db.query.user({
-      where: {
-        id: ctx.request.userId
-      }
-    });
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
     // 3. Check if they have permissions to do this
     hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
     // 4. Update the permissions
     return ctx.db.mutation.updateUser(
       {
-        data: { permissions: { set: args.updatedPermissions } },
+        data: { permissions: { set: args.permissions } },
         where: { id: args.userId }
       },
       info
